requirements.txt
# Core ML and Data Processing (Python 3.11 optimized)
tensorflow>=2.13.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Medical Image Processing
Pillow>=10.0.0
opencv-python>=4.8.0
scikit-image>=0.21.0
pydicom>=2.4.0  # For DICOM medical images
SimpleITK>=2.3.0  # Medical image processing

# Deep Learning for Medical Images
keras>=2.13.0

# Web Interface for Healthcare Professionals
streamlit>=1.28.0
streamlit-option-menu>=0.3.6
streamlit-aggrid>=0.3.4

# Google Gemini AI (Latest)
google-generativeai>=0.8.0

# Professional Visualization & Reporting
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# Medical Data Handling & Utilities
python-dotenv>=1.0.0
requests>=2.31.0

# Progress Tracking & Analytics
# sqlite3 and asyncio are built-in Python modules

# Development & Testing
pytest>=7.4.0
jupyter>=1.0.0